v3.1.1 / TODO #239: range-check the PEM fields that size allocations
ring_sig_load took the ring size k and the round count straight off the wire
and multiplied them into blen and stern_ring_alloc's size.  The caller's k is
only compared against the signature's after the reader returns, so
ring_load_members' RING_MAX_K bound never reached it.  Both are now rejected
outside [2, RING_MAX_K] and [1, SDF_MAX_ROUNDS] before either enters an
arithmetic expression, a payload longer than (k, rounds) describe is rejected
rather than truncated, and stern_ring_alloc computes k * rounds in size_t
instead of int (signed overflow before it reached malloc).

The sweep of the file's other parse_be_uint readers that #239 asked for found
a strictly worse instance of the same shape.  Item[5] of an HPKE-Stern-KEM
private key is the QC-MDPC row weight d, and it sizes der_right_align's writes
into two QCMDPC_D*2-byte STACK buffers at three call sites (pkey, kex
--our-kem, dec).  Nothing bounded it: under ASan, d=16 is a 32-byte
stack-buffer-overflow WRITE and d=2^31-1 a 4 GB one.  That is memory
corruption from a parsed file, not the abort #239 was filed for.

Also bounded: the XMSS height and signature depth, which size `1 << h` leaves
and the num_leaves*KEYBYTES allocation behind them (`1 << h` is undefined for
h >= 32), and the HSKE-NL-V2-Duplex declared ciphertext length.

test_weak_key_rejection.sh grows 19 assertions (6 -> 25), each malformed case
paired with a control asserting the untouched artifact still verifies.  The
sanitizers CI job now runs that script too: the class it covers is memory
safety, which a plain build's exit status does not report.

+++ herradura.h
@@ -2073,9 +2073,13 @@ typedef struct {
     BitArray *resp_a, *resp_b;
 } SternRingSig;
 
+/* sz is computed in size_t, not int: k and rounds reach this function from a
+ * PEM reader, and `k * rounds` in int is signed overflow — undefined behaviour
+ * — before the product ever reaches malloc (TODO #239).  Callers must still
+ * range-check their inputs; this only removes the UB behind them. */
 static void stern_ring_alloc(SternRingSig *sig, int k, int rounds)
 {
-    int sz = k * rounds;
+    size_t sz = (size_t)k * (size_t)rounds;
     sig->k      = k;
     sig->rounds = rounds;
     sig->c0     = (BitArray *)malloc(sz * sizeof(BitArray));
